code review nits and minor updates
