--- kafka-client/src/main/java/dev/responsive/kafka/store/StoreUtils.java
@@ -1,22 +1,0 @@
-package dev.responsive.kafka.store;
-
-import java.time.Duration;
-
-public class StoreUtils {
-  /**
-   * Validate and convert the {@link Duration} to milliseconds
-   */
-  public static long durationToMillis(final Duration duration, final String parameterName) {
-    final String errorMsgPrefix =
-        String.format("Cannot convert %s Duration to milliseconds", parameterName);
-    try {
-      if (duration == null) {
-        throw new IllegalArgumentException(errorMsgPrefix + " due to parameter being null");
-      }
-
-      return duration.toMillis();
-    } catch (final ArithmeticException e) {
-      throw new IllegalArgumentException(errorMsgPrefix + " due to arithmetic exception", e);
-    }
-  }
-}
